BANK_ACCOUNTS-splash: Update bank account form to use new LabledInput component
