+++ src/scss/blocks/atom/_button.scss
@@ -2,7 +2,8 @@
   text-align: center;
   font-weight: bold;
   border-radius: 2em;
-  padding: 1em;
+	padding: 1em;
+	margin: 2em;
 	border: solid 1px $dm__btn--outline;
 	height: 4em;
