+++ src/tailwind.scss
@@ -57146,6 +57146,7 @@ label {
   font-weight: bold;
   border-radius: 2em;
   padding: 1em;
+  margin: 2em;
   border: solid 1px $dm__btn--outline;
   height: 4em;
   &__graphControl {
