+++ src/views/accounts/Accounts.jsx
@@ -6,11 +6,11 @@ import NewBankAccount from '../../components/forms/bankAccount/NewBankAccount';
 const Accounts = () => {
 
 	return (
-		<section>
+		
 			<Card>
 				<NewBankAccount />
 			</Card>
-		</section>
+		
 	);
 }
